(chore): Rename auth to userAuth.
